requirements.txt
chromadb==0.3.29
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
openai
pypdf2
python-dotenv
tiktoken
